requirements.txt
fastapi
uvicorn
requests
aiohttp
boto3
pyjwt
websockets